moving go mod files
