+++ di/di-utils/types_test.go
@@ -26,8 +26,8 @@ func TestTypeOf_Alias(t *testing.T) {
 
 func TestNameOf_CustomType(t *testing.T) {
 	got := NameOf[sample]()
-	if got != "github.com/lcrux/go-di/di-utils/sample" {
-		t.Fatalf("expected github.com/lcrux/go-di/di-utils/sample, got %s", got)
+	if got != "github.com/lcrux/go-di/di/di-utils/sample" {
+		t.Fatalf("expected github.com/lcrux/go-di/di/di-utils/sample, got %s", got)
 	}
 }
